requirements.txt
djangorestframework
line-bot-sdk
gunicorn
heroku